requirements.txt
Django==1.8.2
dj-database-url==0.3.0
psycopg2==2.6
djangorestframework==3.1.2
django-filter==0.10.0
django-cors-headers==1.1.0
djangorestframework-jwt==1.5.0
Markdown==2.6.2